+++ code/builders/builder-vite/src/plugins/storybook-project-annotations-plugin.ts
@@ -5,8 +5,8 @@ import type { Plugin } from 'vite';
 import { generateProjectAnnotationsCode } from '../codegen-project-annotations';
 import { getResolvedVirtualModuleId } from '../virtual-file-names';
 
-const VIRTUAL_ID = 'virtual:/@storybook/builder-vite/project-annotations.js';
-export const RESOLVED_VIRTUAL_ID = getResolvedVirtualModuleId(VIRTUAL_ID);
+export const VIRTUAL_ID = 'virtual:/@storybook/builder-vite/project-annotations.js';
+const RESOLVED_VIRTUAL_ID = getResolvedVirtualModuleId(VIRTUAL_ID);
 
 /**
  * A Vite plugin that serves the project annotations virtual module.
@@ -33,19 +33,7 @@ export function storybookProjectAnnotationsPlugin(options: Options): Plugin {
     },
     async load(id) {
       if (id === RESOLVED_VIRTUAL_ID) {
-        const code = await generateProjectAnnotationsCode(options, projectRoot);
-
-        const hmrCode = [
-          'if (import.meta.hot) {',
-          '  import.meta.hot.accept((newModule) => {',
-          '    window.__STORYBOOK_PREVIEW__?.onGetProjectAnnotationsChanged({',
-          '      getProjectAnnotations: newModule.getProjectAnnotations,',
-          '    });',
-          '  });',
-          '}',
-        ].join('\n');
-
-        return `${code}\n\n${hmrCode}`;
+        return generateProjectAnnotationsCode(options, projectRoot);
       }
     },
   };
